Update dataTables.responsive.js
I use DataTables with scroller and responsive extension.

If scrolling.sY or scrolling.sX are not null, widthAvailable is substract by a scrollbar width even if there isn"t enough data to scroll (and no scrollbar display).
So widthAvailable>usedWidth is false and put the datatable on responsive state.

I did my test with sScrollBodyDom.scrollHeight>sScrollBodyDom.clientHeight where sScrollBodyDom is the div dataTables_scrollBody. I don't know if it is good but it is enough to fix my problem.